Removed the AddCustomer form, and combined the customer and order menu models into one
