--- FoodOrderingApp/Forms/AddCustomerForm.cs
@@ -1,59 +1,0 @@
-﻿using FoodOrderingApp.DB_Data;
-using FoodOrderingApp.Models;
-using System;
-using System.Collections.Generic;
-using System.ComponentModel;
-using System.Data;
-using System.Drawing;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.Windows.Forms;
-
-namespace FoodOrderingApp
-{
-    public partial class AddCustomerForm : Form
-    {
-        public AddCustomerForm()
-        {
-            InitializeComponent();
-        }
-
-        private void Form2_Load(object sender, EventArgs e)
-        {
-
-        }
-
-        private void btnBack2_Click(object sender, EventArgs e)
-        {
-            DialogResult choice =
-               MessageBox.Show("Are you sure you want to Exit?", "Attention",
-                   MessageBoxButtons.YesNo,
-                       MessageBoxIcon.Warning);
-            if (choice == DialogResult.Yes)
-            {
-                this.Close();
-            }
-        }
-        
-        private void btnAddCustomer_Click(object sender, EventArgs e)
-        {
-            var c = new AddCustomer
-            {
-                FirstName = txtFirstName.Text,
-                LastName = txtLastName.Text,
-                PhoneNumber = txtPhoneNumber.Text
-            };
-
-            // Save the customer data to the database
-            FoodOrderingContext context = new();
-
-            context.AddCustomers.Add(c);
-            context.SaveChanges();
-
-            MessageBox.Show("Customer added successfully!");
-            this.Close();
-
-        }
-    }
-}
